A CompoundFileReader and CompoundFileWriter and associated test to show how it works. This will work in the place of Zip to store many files within a single file.
The "directory" must be able to fit in memory in the CompoundFileReader (for performance reasons). To make this store even more files it would need to be modified to read the information as requested (and potentially cache using a MRU list or something like that).

If the file gets corrupted, there is currently no way to recover, just delete and re-create.

git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/icb-commons/icb-utils@15786 60cd2ccc-2b05-0410-9566-da273a940bff